backend/requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# PDF Processing (from your s3.py)
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1
pandas==2.1.4

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# XML Processing
lxml==5.0.0

# JSON
simplejson==3.19.2

# File handling
python-magic==0.4.27

# Rate Limiting (optional)
Flask-Limiter==3.5.0

# Logging
colorlog==6.8.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Development
autopep8==2.0.4

# OCR Libraries
pytesseract==0.3.10          # Tesseract OCR wrapper
Pillow==10.1.0                # Image processing
pdf2image==1.16.3             # Convert PDF pages to images

# Ollama Client
ollama==0.1.7                 # Ollama Python client



